Updating and consolidating share/trade code
